Resources/function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

adal==1.2.7
adlfs==2021.10.0
aiohttp==3.8.0
aiosignal==1.2.0
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.10
argcomplete==1.12.3
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
azure-appconfiguration==1.1.1
azure-applicationinsights==0.1.0
azure-batch==11.0.0
azure-cli==2.30.0
azure-cli-core==2.30.0
azure-cli-telemetry==1.0.6
azure-common==1.1.27
azure-core==1.20.1
azure-cosmos==3.2.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-functions==1.7.2
azure-graphrbac==0.60.0
azure-identity==1.7.1
azure-keyvault==1.1.0
azure-keyvault-administration==4.0.0b3
azure-keyvault-keys==4.5.0b5
azure-keyvault-secrets==4.3.0
azure-loganalytics==0.1.0
#azure-mgmt==4.0.0
azure-multiapi-storage==0.7.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==12.10.0b1
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-file-datalake==12.6.0b1
azure-storage-queue==1.4.0
azure-synapse==0.1.1
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.8.0
azure-synapse-managedprivateendpoints==0.3.0
azure-synapse-spark==0.2.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
DBUtils==2.0.2
debugpy==1.5.1
decorator==5.1.0
delta==0.4.2
delta-spark==1.0.0
Deprecated==1.2.13
distro==1.6.0
Django==3.2.9
entrypoints==0.3
fabric==2.6.0
frozenlist==1.2.0
fsspec==2021.11.0
humanfriendly==9.2
idna==2.10
image==1.5.33
invoke==1.6.0
ipykernel==6.5.0
ipython==7.29.0
isodate==0.6.0
javaproperties==0.5.2
jedi==0.18.0
jmespath==0.10.0
jsondiff==1.2.0
jupyter-client==7.0.6
jupyter-core==4.9.1
knack==0.8.2
matplotlib-inline==0.1.3
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
multidict==5.2.0
nest-asyncio==1.5.1
numpy==1.21.4
oauthlib==3.1.1
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.0
opencensus-ext-requests==0.7.6
opencensus-ext-threading==0.1.2
opencensus-extension-azure-functions==1.0.0
packaging==20.9
pandas==1.3.4
paramiko==2.8.0
parso==0.8.2
pathlib2==2.3.6
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.0
portalocker==1.7.1
prompt-toolkit==3.0.21
psutil==5.8.0
#py4j==0.10.9.2
py4j==0.10.9
pyarrow==6.0.0
pycparser==2.21
PyGithub==1.55
Pygments==2.10.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.6
#pypiwin32==223
pyreadline==2.1
PySocks==1.7.1
#pyspark==3.2.0
pyspark==3.1.2
python-dateutil==2.8.2
pytz==2019.1
#pywin32==302
PyYAML==6.0
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
scp==0.13.6
semver==2.13.0
six==1.16.0
sqlparse==0.4.2
sshtunnel==0.1.5
tabulate==0.8.9
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==0.56.0
wrapt==1.13.3
xmltodict==0.12.0
yarl==1.7.2